Extract statements_system_variable.go
